requirements.txt
aniso8601==9.0.1
apispec==5.1.1
APScheduler==3.8.0
bidict==0.21.4
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
eventlet==0.32.0
fastcore==1.3.26
Flask==2.0.2
flask-apispec==0.11.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
flup==1.0.3
flup-py3==1.0.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyJWT==2.3.0
pyparsing==3.0.1
pypyodbc==1.3.4
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==5.4.1
pytz==2021.3
requests==2.26.0
six==1.16.0
sockets==1.0.0
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
webargs==8.0.1
Werkzeug==2.0.2
wfastcgi==3.0.0
zipp==3.6.0
